+++ apps/judicial-system/web-e2e/src/integration/Court/courtRecord.spec.ts
